fix: address review feedback for update limit changes

+++ backend/src/main/java/com/ccms/portal/dto/response/CardLimitResponse.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 @Builder
 public class CardLimitResponse {
     private Long cardId;
-    private BigDecimal oldLimit;
-    private BigDecimal newLimit;
-    private BigDecimal availableLimit;
+    private Double oldLimit;
+    private Double newLimit;
+    private Double availableLimit;
 }
